gh-140657: Run test_import single phase init tests in a subprocess
Some tests such as test_basic_multiple_interpreters_main_no_reset()
leak memory. Run them in a subprocess to avoid leaking memory.

Remove @no_rerun decorator, it's no longer needed.
